Battery: resolve last footnote about "weak" charging
The kernel can report the information needed to be able to determine a
"weak" charge, in which the system load is higher than the charge that
is happening, but it can be done with the existing information.  It
should still report that the battery is charging, even if the overall
battery capacity is going down.

This was quite common on early Android devices with small batteries and
high WiFi chip power needs.  The "high power" USB chargers have helped
eliminate this issue.

Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>