Hardware Model: resolve the diagram footnote
We don't need to have yet-another-networ-diagram anymore now that the
MDK is up to date and contains lots of hardware information and diagrams
about what the system looks like.  So refer to that instead of trying to
keep things in sync here.

This removes the static image from the documet as it is no longer
needed.

Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>