Formatting: add the "correct" Google header/footer to the document
As requested by Elwin, this changes the header / footer to match the MDK
header/footer.

We do some tricks to create an 'extra space' character based on a
unicode character, to provide some whitespace in the bottom line of the
footer.

The document is now dated as to when the html and pdf files were created
from the source files, again matching the MDK requirements.

Cc: Elwin Ong <elwino@google.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>