Connection Protocols: convert table to "native" rst type
There was only one table in this section, in the "sphinx" format, so
convert it to "native" rst format.  In this conversion, move some of the
wording of the "meaning" out into the text itself, as a table shouldn't
have multiple sentances in a single field.

Also explicitly define GB_OP_INTERNAL as a value, showing that this is
an internal use value, not just some "invalid" field.

Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>