Greybus Documentation: add back latexpdf build target
It turns out the rst2pdf package doesn't make all that "pretty" of a
.pdf file, unlike LaTeX, so let's move back to the LaTeX interface.
This patch adds the target back, and adds some custom configurations to
have the correct Google footer and some other needed tweaks.

Note, some Unicode characters are not printed out at all (like the
servicemark and degree symbol), that is a font issue, and will be fixed
later when I figure out how to find the correct unicode LaTeX font to
use.

But for now, this should be good enough for the MDK 0.2 release.

Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>