Add various "this is a draft" signals.
Add the following so that the generated PDF is obviously not a
released version, to prevent confusion when emailing drafts around:

- Add a "DRAFT" watermark
- Add "DRAFT - DO NOT REDISTRIBUTE" to header and footer
- Change version and release to '0.1-dirty' in conf.py.

There will be a better way to do this when Greg's monkeypatch branch
lands, but this is hopefully useful for now. We can add logic to not
include these signals for release versions for MDK v0.21.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>