Add table figure support.
This comes from https://github.com/jterrace/sphinxtr which provides a
bunch of tweaks to sphinx to allow for figures to be numbered and
referenced, easier ability to change table layout, and other goodness we
want for our document.

All of that was licened under the Creative Commons Attribution 3.0
License, which makes it ok to add to our repo as well.

Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>