table conversion: Convert a bunch of tables to look "correct"
This uses the figtable command for tables to make them show up as a
numbered table (which was requested by Paul and others), and it allows
us to reference the table in the text as well.

Not all tables are fixed up yet, but this is a good start, more will
come later, but this addresses a number of the chapters.

Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>