Bridged PHY: PWM: delete confusing prescriptive text.
The grammar here is hard to parse, and there is no "busy" error in the
protocol status codes table (table-connectin-status-values). The
closest match is GB_OP_RETRY, but that's definitely wrong.

Just delete the negative consequence for now, essentially leaving it
implementation-defined to decide what to do if an attempt is made to
configure the polarity of an enabled PWM instance.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>