greybus-spec: status is in the header
This is only a start, but the status of a request is no longer
in the response payload, it's in the header for the response.
A bunch of tables and wording in individual operation definitions
still need to be fixed, but this fixes the wording in some of
the generic text.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>