greybus-spec: fix some funky characters
In device_class.rst there are symbols representing "1/10 degree
celcius" that show up as strange characters in the text source, and
don't properly format in the PDF output.

Replace the "1/10" with "0.1".  Fix the degree symbol by defining a
macro to represent "degrees celcius."  That *still* doesn't show up
right in the PDF output, but isolates the affected text to
facilitate a future fix.

Similarly, define and use a macro for Greek letter mu (representing
"micro").

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>