greybus-spec: status is in the response header
At one time the status byte was part of the payload of every
response message.  Long ago (seemingly) it was moved into the header
instead.  Update the descriptions of response message payloads to
reflect this.  Use the same basic wording for no-payload response
messages as is used for requests.  Remove the status byte from the
tables defining response message structures.

Signed-off-by: Alex Elder <elder@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>