greybus-spec: fix version operations
A version operation needs to be a negotiation between modules, not a
simple query.  Otherwise if an old AP is used with a module running
newer version protocols, the AP may be unable properly handle some
of the operations supported by the module.

All protocols need to do this negotiation, so provide a generic
high-level description of how this process works in the general
description of protocols and operations.

The patches that follow will update the rest of the spec (in steps).

Signed-off-by: Alex Elder <elder@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>