greybus-spec: update version function prototypes
Continuing with updating the way the version operation is described,
every protocol provides a "conceptual" description of the version
operation in the form of a C function prototype.  Replace these for
all protocols so they use a common definition and explanation that
matches the way it's really done.  Rename the operation "version".

Signed-off-by: Alex Elder <elder@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>