greybus-spec: define protocol data requirements
There is a single "general requirements" section before the
definition of the Manifest.  Most of the requirements apply
to messages as well, but the placement of the requirements
in the document makes things a little less than clear.

Add a new section, "Message Data Requirements" that states the
requirements that apply to messages.  State clearly that alignment
of fields within a message is not required, and that if needed, it
must be achieved through the definition of explicit reserved fields.

Reorder and slightly reword the requirements above the manifest
description.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>