greybus-spec: add explicit padding to descriptors
We've been lax about ensuring the descriptor structures in the
manifest adhere to the alignment rules we set out for them.  This
patch enforces the rule that structures are aligned to four-byte
boundaries and padded to be a multiple of four bytes in size.
It also updates a few size fields as a result.

Where needed, pad fields are explicitly added so the structure
definitions are unambiguous about them.

This does not affect message structures, which don't explicitly
require alignment.  There may be reasons to have explicit padding in
some message structures (I2S has a few of these), but that's not
done here.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>