greybus-spec: define special control protocol
Create a new section of the document to hold the definition of the
control protocol and the SVC protocol.  These two protocols are
special in the operation of a Greybus system.

Populate this "special" section with the definition of the control
protocol, which consists of three operations:  probe; connected; and
disconnected.  The control protocol uses a connection involving the
reserved control CPort on an interface (which all interfaces must
implement).  The "probe" operation is initiated only by the SVC.
The "connected" and "disconnected" messages may be initiated only by
the SVC or the AP.

Normally, an interface is notified (via a "connected" operation)
that one of its CPorts is connected and usable (for both sending
and receiving operation messages).  This is not the case for the
control CPort--the interface will not be notified when it is
connected, and must be ready to receive an operation request message
at any time.  Furthermore, an interface cannot send operation
request messages using its control CPort; a usable connection can
only be assumed for the purpose of responding to an incoming
request.

The other special protocol is the SVC protocol (whose definition
will be added soon).  The SVC protocol is used between the AP and
the SVC for Greybus system management.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>