greybus-spec: replace defines.rst with prologue and epilogue
It's annoying and error-prone to include our substitution definitions
in each new source file. Instead of that, use the special rst_prolog
and rst_epilog values to define strings that will be included at the
beginning and end of every .rst file found in a toctree.

For discoverability, load those strings from files instead of defining
them directly in conf.py.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>