greybus-spec: add Bibliography
For now, this just contains the UniPro spec itself. A later patch will
use this citation.

Note that this appears in a chapter named "Bibliography" in the LaTeX
output, but we don't actually specify that. The LaTeX translator seems
to want to collect all references throughout the document and
consolidate them in a section at the end under that title.

By convention, we will add all new entries here for readability of the
Specification's source code.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>