greybus-spec: define SVC protocol operations
This patch defines the initial set of operations for the SVC
protocol, which exists only on a single connection between the SVC
and the AP Module.  The SVC sets up this connection, and notifies
the AP of its existence through a "connected" control operation.
Thereafter all control of the UniPro network is done under direction
of the AP.

There will definitely be more SVC operations, but this is enough
to get us going with UniPro boostrap.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>