greybus-spec: avoid backslash-escaped whitespace
I added the "\ " sequence following most of the citations for the HID
specification, and they weren't necessary.  So get rid of them.

There was also another place where the table reference numbers were
wrong, so fix that by moving the table.  (Well, the patch shows
the text referencing the table...)

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>