greybus-spec: Bibliography: convert tabs to spaces.
I know, I know. But:

"As in Python, indentation is significant in reST, so all lines of the
same paragraph must be left-aligned to the same level of
indentation" (http://sphinx-doc.org/rest.html).

Given that, it seems better to have all spaces for consistency and so
we don't get puzzling "unexpected unindent" when mixing all-spaces
versus mixed-tabs-and-spaces, since there are no other tabs in any
other ReST files right now.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>