module_info: fix description of String Descriptor padding
Remove incorrect description of String Descriptor padding, which claimed
that the strings themselves rather than the descriptors needed to be
padded to four-byte boundaries, something which would have resulted in
descriptor sizes not divisible by four.

That String Descriptors, like all other descriptors, need to be padded
to make their sizes divisible by four does not have to mentioned
specifically in the describing text. It is also already apparent in the
String Descriptor table.

Signed-off-by: Johan Hovold <johan@hovoldconsulting.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>