greybus-spec: Rename control protocol's probe() operation as probe_ap()
Probe() operation is used only by the SVC at boot time to figure out
which module or interface contains an AP. Only the AP will respond with
a block of authentication data and all non-AP modules respond with no
data.

The current name of this routine 'probe' doesn't clearly communicate its
purpose. It might be better if we call it 'probe_ap()' instead, as
that's what it is supposed to do.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>