greybus-spec: Add get_manifest_size() control protocol operation
Once the AP is discovered by the SVC, the AP can start probing non-AP
modules to get their manifest (which describes the functionality module
provides). Before doing that the AP needs to know the size of the
manifest, to allocate its buffers.

Add another operation to control protocol to get this information.

Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>