greybus-spec/uart: Move parity, framing, overrun, break to send-data message
This patch adds a flags field to the receive-data message and moves parity,
framing, over-run and break reporting to the new field.

If a parity, framing, over-run or break is detected in the Module or gbsim
then the flags field will contain a bitmask of the relevant events being
reported. This patch introduces a mandate that a receive-data message must
have a size field greater than or equal to one byte, with firmware
inserting a NUL byte for break events if required. It also amends the
definition of the receive_data function to add a pass-by-value 'flags'
argument while updating the 'size' argument to similarly be passed by value.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>