greybus-spec/uart: Rationalize all UART flags fields to 8 bits
A flags field has been added to the UART Receive Data message which moves
LSR reporting to the UART Receive Data message. This flags field is 8 bits
wide.

This patch reduces the size of similar fields to 8 bits throughout the UART
protocol for structures encapsulating similar information. It also fixes
the definition of serial_state to ensure to specify the argument is passed
by value.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>