greybus-spec/control: Formally define Control Protocol reserved addresses
Section 9.1 of the greybus specification calls out the requirement for all
interfaces to define a CPort that uses the Control Protocol.
This patch mandates a formal CPort id of '2' for the Control Protocol
descriptor and similarly mandates a formal Bundle id of '0' to be
associated with that CPort descriptor.

This is a desirable change since it formalizes in the specification a
defacto case being applied throughout the development of various greybus
implementations in AP, gbsim and firmware.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>