greybus-spec/uart: make receive-data and serial-state unidirectional
a267770dea31 ('add unidirectional operations') made it possible to specify
some operations as unidirectional operations i.e. operations that do not
require a response. The guidance for when to specify an operation as
unidirectional is to evaluate on a case-by-case basis depending on the
type of Module and the operation in question and decide if that operation
ought to be unidirectional or bidirectional.

Both UART receive-data and serial-state operations are clearly operations
that have no meaningful use for a response from the AP. It is up to the
Unipro bus to guarantee and flow control data transmission to the AP from
a Module and in this context requiring an explicit ACK from the AP is
redundant.

This patch moves both receive-data and serial-state to unidirectional
operation types.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>