sphinxtr: fix figtable multiple definition warnings, wrong numbering
The 'figtable' directive doesn't seem to like being followed directly by
a new section. For some reason, an extra \label is applied to the next
section if figtable is not followed by another paragragh.

This was resulting in hundreds of multiple definition LaTeX warnings, as
well as resulting in incorrect table references, as it applied to both
the table and the immediate following section.

An RST empty comment directive '..' terminates the figtable directive,
fixes the warnings, and the correct table references are now generated

Signed-off-by: Perry Hung <perry@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>