loopback: match transfer request/response payloads
The Transfer Request currently has a 4-byte length field, while the
Response does not. This adds unpleasantness to throughput calculations
based on traffic using this operation.

To make the math easier, as it's one of the main use cases for the
operation, match the request and response payloads exactly by adding a
length field to the response.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Robert Johnson <gtgprods@gmail.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>