greybus-spec/timesync: add feature bit-field to Interface descriptor
This patch adds a feature bit-field to the Interface descriptor
specification. The previously reserved space after the the type field in
the Interface descriptor is re-purposed for the new bit-field. This
bit-field contains just the one defined bit in this patch
GB_INTERFACE_TIME_SYNC. When true GB_INTERFACE_TIME_SYNC indicates that a
Greybus Interface supports time-sync operations.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>