greybus-spec/timesync: add Control Special Protocol timesync extensions
This patch extends the Control Special Protocol to include a set of
timesync operations pertinent to an Interface. This patchset is the logical
Interface implementation of the SVC logic, where SVC is clock master, in
this case an Interface is a clock-slave.

The following operations are added in this patchset:
- timesync_enable
  This operation originates in AP and informs an Interface of an upcoming
  set of TIME_SYNC strobes, with the expected delay given in milliseconds
  between each strobe.

- timesync_disable
  This operation originates in AP and informs an Interface to disable
  frame-time tracking. An Interface should immediately stop the relevant
  counter associated with the frame-time.

- timesync_authoritative
  This operation originates in AP and informs an Interface of the
  authoritative frame-time for-each TIME_SYNC strobe as well as the maximum
  jitter in nanoseconds an Interface should accept when calculating
  propogation delay.

These three operations allow AP driver/power-management software to
enable/disable and update frame-time on an Interface aligned to the SVCs
master clock reference and for the Interface to learn the apparent
propogation delay and adjust the given authoritative frame-time
accordingly.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>