greybus-spec: define a unified "protocol version" operation
Every protocol is copy/pasting similar content for the Protocol Version
operation. This makes it cumbersome to change the meaning of the
operation.

Additionally, different protocols have specified slightly different
semantics for this operation. In some protocols, it's clearly stated
that this operation must be originated by the AP. But in others, it's
left ambiguous, which leaves open the possibility that non-AP interfaces
might have to support backwards compatibility for Greybus protocols.

As prep work to fix all that, add a section declaring the shared
semantics, names, and request/response values for all of these
operations. Later patches will clean up the existing uses to refer to
this section.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>