greybus-spec: add appendix for hotplug
We need a place to store information about how hotplug works in our
system. This has a lot of impact on Greybus, but doesn't necessarily
belong in the core specification itself, since it includes a lot of
implementation details about Ara.

To balance these two, create an appendix file which will contain this
information, but mark it informative.

Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>