SW-2944: build: fix a package conflict
The build system configuration puts the path to the local python
extensions directory at the end of sys.path. That makes python import
mechanism search all the other sys.path directories first before
entering source/extensions.

If a package called backports already exists on the system (as it does
on Debian 8) it is imported and the build fails due to the OrderedDict
class not being present inside it.

This patch prioritises source/extensions over any other directory in
sys.path and fixes the issue.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Tested-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>