svc: fix connection-create flags
Fix the definitions of the SVC connection-create flags that map directly
to UniPro CPort flags.

Our current implementation uses the UniPro definitions for these flag,
whereas the specification so far has defined inverted meanings for CSD_N
and CSV_N.

Instead of requiring another flag-day change of the Greybus and firmware
implementations of this request, simply update the specification to match
the implementation and UniPro spec.

Note that this also has the benefit of not requiring translation between
differing Greybus and UniPro definitions of these flags, something which
facilitates debugging.

Signed-off-by: Johan Hovold <johan@hovoldconsulting.com>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>