greybus-spec: Add reference to the protocols for easier navigation
The top section of special protocols page talks about the protocols that
are going to be defined later.  Navigation on the generated html or pdf
isn't that easy, specially if you want to reach to beginning of a
particular protocol.

Instead of plain text, use references to the individual protocols to
fix that.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>