Classes: Reserve the bridged phy number, and remove specific phy classes
The bridged phy class should encompass all bridged phy "types" so get a
number for this class and mark the existing ones a "reserved" so that
things don't instantly break.

This is part of a larger process of making all protocols be unique per
class type, not global for the whole specification.

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>