greybus-spec: glossary: revamp and refine the definitions
It's been a long time since this section was kept up to date. Let's do
that. We'll need it kept current for some upcoming hotplug-related
changes, anyway.

Most of these are simple refinements of what was already
there. The major breaking changes are:

- "Endo" is now "Frame"
- The "Interface Block" is on the "Frame", not on the "Module"

The first of these is a simple naming update.

The second makes the spec reflect widespread usage, to avoid
confusion. (We usually refer to an "interface" in context of what the
SVC is doing or thinking, which generally means a Frame-centric way of
thinking).

Both will require some propagation throughout the rest of the spec, to
come in further patches.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>