SW-2544: greybus-spec: SVC protocol: define SVC power down operation
Add the "SVC power down" Greybus operation to the SVC Protocol.
This operation is necessary to allow the AP request the SVC to power
down (forcibly) all devices under its control, as well as putting
itself in power down mode.
This shall be used when the phone is being powered down for example.

Signed-off-by: Patrick Titiano <ptitiano@baylibre.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>