greybus-spec: rewrite hardware model
The current hardware model just says "go read the MDK". That's
problematic in that the confusion and imprecision of mapping Greybus
operations to all the different hardware we've historically supported is
leading to conflicting and unstable implementations.

We need to be able to define things in the spec in a way that maps to
the hardware. This needs to be precise enough to cover existing
hardware, but not paint us into a corner when we need to support new
kinds of modules.

This patch provides that. It includes a definition for an Interface
State, which represents the state that the Frame (both AP and SVC)
maintain for each Interface Block. It also defines its initial value,
and leaves hooks for the dynamics of the states within the system.

We'll build on this to define a module lifecycle, as well as the runtime
behavior of the Interfaces, as we go.

Some particular use-cases whose Greybus behavior we want to define in
terms of this model are:

- module "mode switching" and the multi-stage firmware boot process in
  general ("working" -> "crashed", stage 1 -> stage 2 via boot over
  UniPro, and stage 2 -> stage 3 via BoU and authentication to a stage 3
  already present on the module)

- error handling when the link is lost

- power supply modules (i.e. modules that supply power to the frame, not
  power supply to modules from the frame)

- firmware update (e.g. managing the switch resources properly after
  update is complete and synchronizing CPort closure)

- power management (suspend/resume, off/on, etc.)

- reusing CPorts dynamically (including "offloaded" I2S/camera CPorts
  and various combinations of CPort flags)

Some of these use cases will require additional refinement to this
model, but it's a starting point and will serve to let us define the
basic machinery we need for reliable hotplug.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>