control: add bundle-version operation
Add a new Bundle Version operation that is used to retrieve the version
of the bundle class implemented by a bundle.

This is part of the new protocol version handling in which classes have
versions. Generally, the version of a subprotocol of a class is
identical to the version of the bundle class itself (e.g. we do not
version an Audio bundle's two subprotocols independently).

The individual protocol's version-request operations will eventually be
removed (except for Control, SVC, and Firmware which will still need
them).

Signed-off-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>