greybus-spec: hardware-model: move "DETECT" up
The DETECT sub-state really needs to be understood first, before any of
the others, as that's what indicates whether a module is attached to an
interface block or not.

Make that happen by moving the description for the sub-state to the top,
rewriting any fully specified interface states so DETECT comes first,
and updating some other miscellaneous text that depended on the order of
the sub-states.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Alex Elder <elder@linaro.org>
Reviewed-by: Jeffrey Carlyle <jcarlyle@google.com>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>