greybus-spec: hardware_model: s/UPRO_LOST/UPRO_RELINK/
The UPRO_LOST state is really meant to capture the situation when the
peer is re-initiating link startup after a link is already
established. The link was thus "lost" at some unspecified point in the
past, and so UPRO_LOST is a bad name.

Rename the value UPRO_RELINK, and move it after UPRO_UP in the table of
UNIPRO sub-state values to reflect its natural order in the progression
of changes to an Interface State (UNIPRO must be UPRO_UP before it can
be UPRO_RELINK).

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Alex Elder <elder@linaro.org>
Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>