greybus-spec: generalize ping request
Define the ping request to be valid on all connections rather than
just on the control connection.  Use operation type value 0x00 for
this operation.

The SVC protocol uses this operation occasionally as a way to
determine whether the UniPro network is still operational.

In addition, this operation is used in the teardown sequence for all
connections.  The ping operation sends a message in each direction
across a connection, and the request and response are used to signal
the "last message" from a sender.

Signed-off-by: Alex Elder <elder@linaro.org>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>