greybus-spec: svc: add connection quiescing operation
Add a new operation to the SVC protocol that states that a
connection is in its quiescing phase prior to being disconnected.
This gives the SVC a chance to prepare the two connection endpoints
for disconnection.  After this, one final round-trip ping operation
will be performed on the subject connection before the AP requests
it be destroyed.

Signed-off-by: Alex Elder <elder@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>