greybus-spec: svc: new module operations
Define the Module insertion and removal operations.  The first is
used by the SVC to inform the AP that a new Module has been inserted
into the Frame.  It sends this request after it has determined the
location of the Module, which is defined as the position (or id) of
its Primary Interface.  It also supplies the Module's "size", which
is expressed in terms of the number of Interfaces the Module covers.

The second is used by the SVC to tell the AP that a previously-
inserted Module is no longer present.  This can occur as a result
of an eject operation, or could occur in the event a Module is
"forcefully" (unintentionally) removed from the Frame.

Signed-off-by: Alex Elder <elder@linaro.org>
Reviewed-by: Marti Bolivar <mbolivar@leaflabs.com>
Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>