SW-1996: Add TimeSync SVC wd-pins-init/wd-pins-fini
The SVC must establish an initial state on the wake-detect pins for all
relevant timesync participants prior to generating any timesync releated
strobes on wake-detect pins. This ensures that a floating pin when
configured as an output won't be mis-interpreted as a valid state
transition by the downstream timesync participant.

To enable this behaviour a wd_pins_init and wd_pins_fini operation is
required from AP to SVC. This establishing a known state on all output pins
which allows the APB/GPB slaves to be enabled for timesync and to be sure
then that the state-transitions observed at APB/GPB are not as a result of
floating GPIOs being configured as outputs by SVC.

This specification update captures the operations formally.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>