SW-1996: Add TimeSync SVC Ping
In order to verify the frame-time at each timesync participant at arbitrary
deliniations we need the SVC to generate a single timesync event to a known
set of timesync participants. The SVC will return the authoritative
frame-time to the AP at the time of the wake-detect event.

The AP will then poll each participant in the system for the
last-event-time and can present this information, or take further action
based on this data.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>