greybus-spec: specify UNIPRO, MAILBOX, INTF_TYPE relationship
When an Interface is ACTIVATED, we determine three interrelated
sub-states in the Interface State:

- UNIPRO: based on whether the link startup sequence succeeds
- MAILBOX: if link startup sequence succeeds, whether the module
  successfully boots the interface and writes the mailbox
- INTF_TYPE: the interface type, based on a combination of UNIPRO and
  MAILBOX

Specify this relationship in the section on ACTIVATED. This can be
described in more detail later when we describe what happens in the
"boot" edge from ATTACHED -> ACTIVATED in the lifecycle state machine.

Note that other combinations are intentionally not defined. We're
focusing on what's well defined within this section. We will describe
error handling in the upcoming section which defines sequence diagrams
for the lifecycle state transitions -- and errors determining the
interface type are fatal anyway.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>