greybus-spec: add mscgen dependency
Add a dependency on mscgen (a Message Sequence Chart GENerator):

http://www.mcternan.me.uk/mscgen/

It will be used in later patch series to include sequence diagrams in
the specification. If it's premature to include the additional
dependency now since none of the patches in this series depend on it,
this patch can be safely dropped.

Teach the build system that .msc files in source/img/msc/ need to be
converted to .pngs with msgcen so they can be included from .rst files.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>