greybus-spec: define SVC protocol-specific status values
We've never used it, but the Greybus spec allows individual protocols to
define some of their own status bytes.

This is convenient for the SVC protocol, which in some cases needs to
have error handling related to Interface States grafted onto it after
the fact, but for which we'd like to avoid changing any request or
response payload structures if possible to minimize code churn.

To enable that, add some SVC protocol-specific status values that will
let the SVC signal errors when the AP is making requests for Interfaces
whose Interface States aren't "right" for the request in some ways. We
can extend these if we need more.

Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>