greybus-spec: svc: add missing flags to Module Inserted op
At hotplug camp, we designed the module inserted operation to also
contain a flags field, which is missing from the current spec.

We need this flag to handle hardware errors in the module geometry
detection circuits, which prevent the SVC from successfully detecting
which interface to the module is primary.

Add the missing field and its current (only) user.

Reviewed-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>