greybus-spec: add outline of lifecycle transitions chapter
After this point in the series, we will start modifying operation
definitions to refer to the hardware model, and describing how they
affect interface states (or don't).

That's good, but it's also a giant wall of text that will be hard to
understand, implement, or design tests for. We need diagrams and
explanations showing how the operations shall be sequenced to manage the
module and interface lifecycles (and who knows; maybe we'll get around
to writing the bundle lifecycle we really ought to have someday...).

Add a new chapter with the outline of that, i.e. section headers for
what needs explaining, along with a few introductory sections. Inserting
this "place-holder" patch at this point in the series lets us use the
section header ref targets in subsequent patches.

To prevent drift between this chapter and hardware_model, move the
definitions of the Lifecycle States as groups of tuples into text files,
and include those from both files, keeping a single point of truth.

Acked-by: Johan Hovold <johan@hovoldconsulting.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>